modified to run without compile
I cannot compile default setting. So I modified the code using bash
shell script and root macro.
