--- picoNpeProductionMonitor/GNUmakefile
@@ -1,23 +1,0 @@
-OBJS = plotPicoNpeProductionQa.o
-EXE = plotPicoNpeProductionQa
-
-ROOTCFLAGS    = $(shell root-config --cflags)
-ROOTLIBS      = $(shell root-config --libs)
-ROOTGLIBS     = $(shell root-config --glibs)
-
-INCFLAGS = -I$(ROOTSYS)/include
-LDFLAGS = -L$(ROOTSYS)/lib
-
-CXX = g++ -std=c++11 -m64
-FLAGS = -Wall -g $(INCFLAGS) $(LDFLAGS)
-
-COMPILE = $(CXX) $(FLAGS) -c 
-
-all: $(EXE)
-
-$(EXE): $(OBJS)
-	$(CXX) -o $(EXE) $(OBJS) $(ROOTFLAGS) $(ROOTLIBS)
-
-%.o: %.cxx
-	$(COMPILE)  $< 
-
