+++ picoNpeProductionMonitor/README.md
@@ -1,0 +1,2 @@
+Requirement : pandoc, root
+How to use : ./createHtml.sh ROOT_FILE_NAME
