--- picoNpeProductionMonitor/createHtmlFromMd.sh
@@ -1,1 +1,0 @@
-pandoc --standalone --mathjax --from markdown --to html -o index.html index.md
