+++ picoNpeProductionMonitor/index.html
@@ -6,12 +6,12 @@
   <meta name="generator" content="pandoc" />
   <title></title>
   <style type="text/css">code{white-space: pre;}</style>
-  <script src="http://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" type="text/javascript"></script>
 </head>
 <body>
 <h2 id="picodst-production-monitor">PicoDst production monitor:</h2>
 <p><a href="http://portal.nersc.gov/project/star/mustafa/picoProductionMonitor/">Can be found here.</a></p>
-<h2 id="pico-d0-production-qa">Pico <span class="math">\(D^{0}\)</span> production QA:</h2>
-<p><img src="h1AverageGRefMultInRun.png" /> <img src="h1AverageHftTracksInRun.png" /> <img src="h1AverageKaonsInRun.png" /> <img src="h1AveragePionsInRun.png" /> <img src="h1AverageD0CandidatesInRun.png" /> <img src="h2NKaonsVsNPions.png" /> <img src="h2KaonDcaVsPt.png" /> <img src="h2PionDcaVsPt.png" /> <img src="h2CosThetaVsPt.png" /> <img src="h2DcaDaughtersVsPt.png" /> <img src="h2InvariantMassVsPt.png" /></p>
+<h2 id="pico-npe-production-qa">Pico NPE production QA:</h2>
+<p><img src="h1AverageGRefMultInRun.png" /> <img src="h1AverageHftTracksInRun.png" /> <img src="h1AverageElectronsInRun.png" /> <img src="h1AveragePartnersInRun.png" /> <img src="h1AveragePhECandidatesInRun.png" /> <img src="h2NElectronsVsNPartners.png" /> <img src="h2PairDcaVsPt.png" /> <img src="h2InvariantMassVsPt.png" /> <img src="h2ConversionPosition.png" /> <img src="h2ConversionPosition2.png" /></p>
+<p>Last updated on Mon May 18 03:22:42 KST 2015</p>
 </body>
 </html>
