+++ picoNpeProductionMonitor/index.md
@@ -1,15 +1,15 @@
 ##PicoDst production monitor:  
 [Can be found here.](http://portal.nersc.gov/project/star/mustafa/picoProductionMonitor/)  
 
-##Pico $D^{0}$ production QA:  
+##Pico NPE production QA:  
 ![](h1AverageGRefMultInRun.png)
 ![](h1AverageHftTracksInRun.png)
-![](h1AverageKaonsInRun.png)
-![](h1AveragePionsInRun.png)
-![](h1AverageD0CandidatesInRun.png)
-![](h2NKaonsVsNPions.png)
-![](h2KaonDcaVsPt.png)
-![](h2PionDcaVsPt.png)
-![](h2CosThetaVsPt.png)
-![](h2DcaDaughtersVsPt.png)
+![](h1AverageElectronsInRun.png)
+![](h1AveragePartnersInRun.png)
+![](h1AveragePhECandidatesInRun.png)
+![](h2NElectronsVsNPartners.png)
+![](h2PairDcaVsPt.png)
 ![](h2InvariantMassVsPt.png)
+![](h2ConversionPosition.png)
+![](h2ConversionPosition2.png)
+
